Fix editing an inline node converts it to a block node on save
